% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugit-package.R
\docType{package}
\name{debugit-package}
\alias{debugit}
\alias{debugit-package}
\title{debugit: Debug Helpers}
\description{
Debug helpers.
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
